Removed pngs with clear alpha channels. Image is lost if processed through OpenCV color conversion.
Added loadComparisonImages() which will load from filesystem a set of images that the User's drawing will be compared to.
Changed mComparisonImages QImage vector to a ProcessLayer vector. However, this might be changed back to QImage as we do need to scale the images.
Added x and y range member variables to DrawArea class. These variable are used to keep track of the dimensions of the User's drawing.