Awesome demo build! The application will calculate the best fit image based on the current users image. Give it try! Just keep in mind issues and messy code.
Added compareLayer() that will perform the necessary operations to compare the Comparison Images with the User's image.
Added getComparisonImage() to let mainwindow have an image to place up the screen.
In DrawArea, mComparisonImages is no longer a vector of ProcessLayers. It is now a QImage vector to allow for images to be scaled during compareLayer()compareLayer.
Added mPredicationArea member variable to mainwindow. This QLabel object will display the "predicted" image based on what was returned from compareLayer() and getComparisonImage().
Made qImageToCvMat() a global function that is not attached to ProcessLayer class. This is to allow compareLayer() to convert QImages to the correct cv::Mats.
compareLayer() in mainwindow calls both DrawArea::compareLayer() and DrawArea::getComparisonImage().